Restructured Parser files
