+++ src/parser.rs
@@ -1,3 +1,5 @@
 //! The parser for the project.
 
 pub mod parser;
+pub mod parser_instruction_tokenization;
+pub mod parser_preprocessing;
