+++ src/parser/parser.rs
@@ -1,12 +1,9 @@
-pub mod parser_instruction_tokenization;
-pub mod parser_preprocessing;
-
 use std::fs;
 use std::env;
-use crate::parser::parser::parser_instruction_tokenization::*;
-use crate::parser::parser::parser_instruction_tokenization::ErrorType::*;
-use crate::parser::parser::parser_instruction_tokenization::OperandType::*;
-use crate::parser::parser::parser_preprocessing::*;
+use crate::parser::parser_instruction_tokenization::*;
+use crate::parser::parser_instruction_tokenization::ErrorType::*;
+use crate::parser::parser_instruction_tokenization::OperandType::*;
+use crate::parser::parser_preprocessing::*;
 
 
 fn parser() {
@@ -369,8 +366,8 @@ mod all_tokenization_tests {
     }
 
     mod read_register_tests {
-        use crate::parser::parser::parser_instruction_tokenization::ErrorType::UnrecognizedRegister;
         use crate::parser::parser::read_register;
+        use crate::parser::parser_instruction_tokenization::ErrorType::UnrecognizedRegister;
 
         #[test]
         fn read_register_returns_correct_binary_on_valid_register_name() {
@@ -392,8 +389,7 @@ mod all_tokenization_tests {
     }
 
     mod immediate_tests {
-        use crate::parser::parser::parser_instruction_tokenization::ErrorType::{ImmediateOutOfBounds, NonIntImmediate};
-        use crate::parser::parser::read_immediate;
+        use crate::parser::parser::*;
 
         #[test]
         fn read_immediate_returns_error_on_non_int_string() {
@@ -427,8 +423,8 @@ mod all_tokenization_tests {
     }
 
     mod memory_address_tests {
-        use crate::parser::parser::parser_instruction_tokenization::ErrorType::{ImmediateOutOfBounds, InvalidMemorySyntax, NonIntImmediate, UnrecognizedRegister};
         use crate::parser::parser::read_memory_address;
+        use crate::parser::parser_instruction_tokenization::ErrorType::{ImmediateOutOfBounds, InvalidMemorySyntax, NonIntImmediate, UnrecognizedRegister};
 
         #[test]
         fn missing_open_parenthesis_returns_error() {
